Consolidate a single page to manage resources instead of one specific for each resource
