+++ src/components/resource/manage-resources.css
@@ -1,7 +1,6 @@
 .page-container {
   max-width: 1000px;
   margin: 0 auto;
-
   width: 100%;
   box-sizing: border-box;
 }
@@ -9,24 +8,28 @@
   padding-bottom: calc(30px + 5rem);
 }
 
+/* Encabezado del training */
 .training-card-header {
   display: flex;
   align-items: center;
   gap: 0.75rem;
   font-size: 1rem;
   font-weight: bold;
-  margin-bottom: 0.2rem;
+  margin-bottom: 0.3rem;
+  margin-top: 0.3rem;
   width: 100%;
   box-sizing: border-box;
 }
 
+/* Contenedor de sets */
 .sets-container {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 0.5rem;
   width: 100%;
 }
 
+/* Card del set */
 .set-preview {
   border: 1px solid #ddd;
   border-radius: 10px;
@@ -36,25 +39,29 @@
   box-sizing: border-box;
 }
 
+/* Encabezado del set */
 .set-header {
   display: flex;
   align-items: center;
-  justify-content: flex-start; /* <-- agrega esta línea */
+  justify-content: flex-start;
   gap: 0.5rem;
   font-weight: 600;
-  margin-bottom: 0.5rem;
-  font-size: 1.1rem;
+  margin-bottom: 0px;
+  font-size: 1.05rem;
   cursor: pointer;
   user-select: none;
 }
 
+/* Lista de quizzes dentro de un set */
 .training-quizzes-container {
   display: flex;
   flex-direction: column;
   gap: 0.75rem;
   width: 100%;
+  padding-top: 0.5rem;
 }
 
+/* Card del quiz */
 .quiz-preview {
   background-color: white;
   border-radius: 8px;
@@ -62,15 +69,16 @@
   box-shadow: 0 1px 4px rgba(0, 0, 0, 0.05);
   width: 100%;
   box-sizing: border-box;
+  font-size: small;
 }
 
+/* Chips de secciones */
 .sections-preview {
   margin-top: 0.5rem;
   display: flex;
   flex-wrap: wrap;
   gap: 0.5rem;
 }
-
 .section-chip {
   background-color: #e0f0ff;
   padding: 0.25rem 0.6rem;
@@ -80,6 +88,7 @@
   color: #337ab7;
 }
 
+/* Botón de toggle */
 .toggle-btn {
   font-weight: bold;
   background: none;
@@ -91,46 +100,47 @@
   align-items: center;
 }
 
+/* Badge de rol */
 .role-owner {
   background-color: #007bff;
   color: white;
 }
-
 .role-read-write {
   background-color: #ffc107;
   color: black;
 }
-
 .role-read-only {
   background-color: #6c757d;
   color: white;
 }
-
 .role-unknown {
   background-color: #dc3545;
   color: white;
 }
+
+/* Metainformación del training */
 .training-meta {
-  margin-top: 1rem;
-  font-size: 0.85rem;
-  color: #444;
-  background-color: #fcfcfc;
-  border-top: 1px solid #eee;
-  padding-top: 0.75rem;
+  margin-top: 0.5rem;
+  padding-top: 0.5rem;
+  border-top: 1px solid #e0e0e0;
+  color: #555;
+  font-size: 0.9rem;
   display: flex;
   flex-direction: column;
-  gap: 0.25rem;
+  gap: 4px;
 }
 
-.role-owner {
-  background-color: #007bff;
-  color: white;
+/* Etiqueta sets */
+.sets-label {
+  font-weight: bold;
+  color: #000000;
+  font-size: 1.1rem;
 }
-.role-read-write {
-  background-color: #ffc107;
-  color: black;
-}
-.role-read-only {
-  background-color: #6c757d;
-  color: white;
+.sets-count {
+  background-color: #e0f0ff;
+  color: #007bff;
+  padding: 2px 10px;
+  border-radius: 12px;
+  font-weight: 600;
+  font-size: 0.9rem;
 }
