--- src/components/training/training-card.css
@@ -1,49 +1,0 @@
-.training-card {
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  padding: 1rem;
-  margin: 0.5rem;
-  background-color: #fff;
-  transition: box-shadow 0.2s ease-in-out;
-  cursor: pointer;
-  width: 300px;
-}
-
-.training-card.selected {
-  border-color: #007bff;
-  background-color: #f0f8ff;
-}
-
-.resource-card-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.resource-title {
-  font-size: 1.1rem;
-  font-weight: bold;
-  margin: 0;
-}
-
-.delete-button {
-  background: none;
-  border: none;
-  color: #cc0000;
-  cursor: pointer;
-  font-size: 1rem;
-}
-
-.delete-button:hover {
-  color: #ff0000;
-}
-
-.resource-card-body {
-  margin-top: 0.5rem;
-}
-
-.training-set-preview {
-  font-size: 0.9rem;
-  margin-top: 0.3rem;
-  color: #444;
-}
